move audio alarm to toolbar, and overlayscreen to app
