+++ frontend/src/modules/app/App.tsx
@@ -17,6 +17,7 @@ import { getFrontendDisplaySetting } from '../../store/controller/selectors/fron
 import { StoreState } from '../../store/types';
 import { darkTheme, lightTheme } from '../../styles/customTheme';
 import Routes from './navigation/Routes';
+import OverlayScreen from './OverlayScreen';
 
 const RIGHT_CLICK_BUTTON_KEY = 2;
 const TOUCH_SCREEN_BUTTON_KEY = 0;
@@ -79,6 +80,7 @@ class App extends Component<Props> {
     return (
       <IntlProvider key={locale} locale={locale} messages={messages}>
         <MuiThemeProvider theme={theme}>
+          <OverlayScreen />
           <CssBaseline />
           <Routes />
         </MuiThemeProvider>
